Correcting calculations of uncertainties
Correcting estimation of uncertainties from inverse Hessian. Before, the estimation was done for the first M parameters where
M is number of optimized (i.e. not-fixed) parameters regardless of whether those parameters were fixed or not. In the corrected 
version the Hessian is calculated (estimated) only for the optimized parameters. Additionally, to calculated the uncertainties, the 
inv(Hessian)*chi2 should be normalized by square root of degrees off freedom (DOF = number of data points - number of fitted parameters) of the fit , a corresponding change was made in inline_estimate_uncertainty.m